Added option to use webhook instead of polling
To enable the webhook mode, the environment variable `WEBHOOK_URL` must be defined with the public URL (it supposes that there is a reverse proxy that enables HTTPS for the public part). By default, expects to receive all the requests without any path prefix, but you can define one using the env var `WEBHOOK_PATH_PREFIX`. Both env vars can contain the token `{token}` that will be replaced with the real token if needed.